requirements.txt
psycopg2_binary==2.9.5
pytest==6.2.5
